requirements.txt
aiodns==3.2.0
aiofiles==0.8.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
bbox-visualizer==0.1.0
beautifulsoup4==4.12.3
brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
gdown==5.2.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
jinja2==3.1.4
kiwisolver==1.4.7
landingai==0.3.48
markdown-it-py==3.0.0
markupsafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
multidict==6.1.0
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==10.4.0
propcache==0.2.0
psutil==6.0.0
pycares==4.4.0
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.5.2
pygments==2.18.0
pyparsing==3.1.4
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
pyyaml==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
segmentation-mask-overlay==0.3.4
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.6
tenacity==8.5.0
tqdm==4.66.5
typer==0.12.5
typing-extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
uvicorn==0.31.1
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
yarl==1.14.0
